structure updates, database updates

+++ Project/src/main/java/Controllers/StudentInfoController.java
@@ -1,10 +1,12 @@
 package Controllers;
 
 import Models.StudentUser;
+import Models.Subject;
 import Models.User;
 import Repository.UserRepository;
-import Services.UserAuthService;
-import Services.WindowSizeUtils;
+import Services.*;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -16,7 +18,7 @@ import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
-import javafx.stage.FileChooser;
+import javafx.scene.layout.VBox;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 
@@ -26,9 +28,10 @@ import java.sql.SQLException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ResourceBundle;
-import java.util.UUID;
 
 public class StudentInfoController implements Initializable {
+    @FXML
+    private VBox subjectsBox;
     @FXML
     private ImageView birthdayw;
     @FXML
@@ -41,6 +44,8 @@ public class StudentInfoController implements Initializable {
     private ImageView emailw;
     @FXML
     private ImageView gradeLvlw;
+    @FXML
+    private ImageView paralelw;
     @FXML
     private ImageView addressw;
     @FXML
@@ -62,6 +67,8 @@ public class StudentInfoController implements Initializable {
     @FXML
     private Spinner<Integer> gradeLvl;
     @FXML
+    private Spinner<Integer> paralel;
+    @FXML
     private DatePicker birthday;
     @FXML
     private TextField email;
@@ -78,33 +85,31 @@ public class StudentInfoController implements Initializable {
     private boolean isEditable = false;
     private String imagePath;
     private String oldImagePath;
+    private StudentUser studentUser;
+    private ObservableList<Subject> subjectsList = FXCollections.observableArrayList();
 
     @FXML
-    private void handleImageUploadButton(ActionEvent event) {
-        FileChooser fileChooser = new FileChooser();
-        fileChooser.setTitle("Choose Image");
-        fileChooser.getExtensionFilters().addAll(
-                new FileChooser.ExtensionFilter("Image Files", "*.png", "*.jpg", "*.gif")
-        );
-        selectedFile = fileChooser.showOpenDialog(primaryStage);
-        if (selectedFile != null) {
-            Image newImage = new Image(selectedFile.toURI().toString());
-            profilePic.setImage(newImage);
-        }
+    private void handleImageUploadButton(ActionEvent event) throws IOException {
+      selectedFile = GeneralUtil.handleImageUpdate(profilePic);
 
 
     }
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
+
         birthdayw.setVisible(false);
         firstnamew.setVisible(false);
         lastnamew.setVisible(false);
         phonew.setVisible(false);
         emailw.setVisible(false);
         gradeLvlw.setVisible(false);
+        paralelw.setVisible(false);
         addressw.setVisible(false);
         w.setVisible(false);
         gradeLvl.setValueFactory(new SpinnerValueFactory.IntegerSpinnerValueFactory(0, 12, 0));
+        gradeLvl.setDisable(true);
+        paralel.setDisable(true);
+        paralel.setValueFactory(new SpinnerValueFactory.IntegerSpinnerValueFactory(0, 3, 0));
         Image oldImage =profilePic.getImage();
         oldImagePath = oldImage.getUrl();
         updatePhoto.setVisible(false);
@@ -150,40 +155,33 @@ public class StudentInfoController implements Initializable {
             address.setEditable(true);
             email.setEditable(true);
             gradeLvl.setDisable(false);
+            paralel.setDisable(false);
             updatePhoto.setVisible(true);
             removeButton.setVisible(true);
             editButton.setText("Save");
             isEditable = true;
         } else {
+            removeButton.setVisible(false);
+            birthdayw.setVisible(false);
+            firstnamew.setVisible(false);
+            lastnamew.setVisible(false);
+            phonew.setVisible(false);
+            emailw.setVisible(false);
+            gradeLvlw.setVisible(false);
+            paralelw.setVisible(false);
+            addressw.setVisible(false);
+            w.setVisible(false);
+            System.out.println("TEst");
 
                 if (selectedFile != null) {
-                    try {
-                        String fileName = selectedFile.getName();
-                        String extension = fileName.substring(fileName.lastIndexOf("."));
-                        String newFileName = UUID.randomUUID().toString() + extension;
-                        imagePath = "src/main/resources/Images/ProfilePics/" + newFileName;
-                        File destination = new File(imagePath);
-                        InputStream source = new FileInputStream(selectedFile);
-                        OutputStream output = new FileOutputStream(destination);
-                        byte[] buffer = new byte[1024];
-                        int length;
-                        while ((length = source.read(buffer)) > 0) {
-                            output.write(buffer, 0, length);
-                        }
-                        source.close();
-                        output.close();
-                        System.out.println("Writing file to " + imagePath);
+                    imagePath = GeneralUtil.savePhoto(selectedFile);
+                    String relativePath = localStudent.getProfileImg();
 
-                        String relativePath = localStudent.getProfileImg();
-//                                .replace("src/main/resources", "");
-                        File oldImage = new File(relativePath);
+                    File oldImage;
+                    if(relativePath!=null){
+                        oldImage = new File(relativePath);
                         oldImage.delete();
-                        System.out.println(oldImage.getAbsolutePath());
-
-                    } catch (IOException ex) {
-                        System.out.println("Error saving image: " + ex.getMessage());
                     }
-
                 }
                 else {
                     imagePath = localStudent.getProfileImg();
@@ -195,7 +193,7 @@ public class StudentInfoController implements Initializable {
                 if(emailValue.matches(emailRegex)){
                     if( !this.firstname.getText().isEmpty() && !this.lastname.getText().isEmpty() &&
                             !(this.birthday.getValue() == null) && !this.phone.getText().isEmpty() &&
-                            !this.address.getText().isEmpty() && this.gradeLvl.getValue() != null){
+                            !this.address.getText().isEmpty() && this.gradeLvl.getValue() != null && this.paralel.getValue() != null){
                             String nameValue = this.firstname.getText();
                             String surnameValue = this.lastname.getText();
                             LocalDate tempBirthdateValue = this.birthday.getValue();
@@ -204,10 +202,11 @@ public class StudentInfoController implements Initializable {
                             String phoneValue = this.phone.getText();
                             String addressValue = this.address.getText();
                             int yearValue = this.gradeLvl.getValue();
+                            int paralelValue = this.paralel.getValue();
                             System.out.println("ALL GOOD");
                             try {
                                 System.out.println("Register -> AuthService");
-                                User user = UserAuthService.updateStudent(student.getID(), nameValue, surnameValue, birthdateValue, phoneValue, addressValue, yearValue, emailValue,"Student", imagePath);
+                                User user = UserAuthService.updateStudent(student.getID(), nameValue, surnameValue, birthdateValue, phoneValue, addressValue, yearValue,paralelValue, emailValue,"Student", imagePath);
                                 firstname.setEditable(false);
                                 lastname.setEditable(false);
                                 phone.setEditable(false);
@@ -215,6 +214,7 @@ public class StudentInfoController implements Initializable {
                                 address.setEditable(false);
                                 email.setEditable(false);
                                 gradeLvl.setDisable(true);
+                                paralel.setDisable(true);
                                 updatePhoto.setVisible(false);
                                 editButton.setText("Edit or Delete Student");
                                 isEditable = false;
@@ -225,9 +225,9 @@ public class StudentInfoController implements Initializable {
                                 phonew.setVisible(false);
                                 emailw.setVisible(false);
                                 gradeLvlw.setVisible(false);
+                                paralelw.setVisible(false);
                                 addressw.setVisible(false);
                                 w.setVisible(false);
-                                System.out.println("TEst");
                             }catch (SQLException sqlException) {
                                 System.out.println("Student couldn't be updated.");
                             }
@@ -253,12 +253,42 @@ public class StudentInfoController implements Initializable {
                         if (this.gradeLvl.getValue() == null) {
                             gradeLvlw.setVisible(true);
                         }
+                        if (this.paralel.getValue() == null) {
+                            paralelw.setVisible(true);
+                        }
+                        if (this.email.getText().isEmpty()){
+                            emailw.setVisible(true);
+                        }
                     }
 
                 }else{
                     System.out.println("Please check your email and try again!");
                     emailw.setVisible(true);
                     w.setVisible(true);
+                    if (this.firstname.getText().isEmpty()) {
+                        firstnamew.setVisible(true);
+                    }
+                    if (this.lastname.getText().isEmpty()) {
+                        lastnamew.setVisible(true);
+                    }
+                    if (this.birthday.getValue() == null) {
+                        birthdayw.setVisible(true);
+                    }
+                    if (this.phone.getText().isEmpty()) {
+                        phonew.setVisible(true);
+                    }
+                    if (this.address.getText().isEmpty()) {
+                        addressw.setVisible(true);
+                    }
+                    if (this.gradeLvl.getValue() == null) {
+                        gradeLvlw.setVisible(true);
+                    }
+                    if (this.paralel.getValue() == null) {
+                        paralelw.setVisible(true);
+                    }
+                    if (this.email.getText().isEmpty()){
+                        emailw.setVisible(true);
+                    }
                 }
 
             }
@@ -266,6 +296,16 @@ public class StudentInfoController implements Initializable {
 
     public void setStudent(StudentUser student) {
         this.student = student;
+        try {
+            subjectsList = FetchData.getStudentSubjects(student);
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+        for (Subject subject:subjectsList
+             ) {
+            System.out.println(subject.getName());
+        }
+        CardGenUtil.subjectsToFlowPaneStudents(subjectsBox, subjectsList, student);
         if (student != null) {
             firstname.setText(student.getName());
             lastname.setText(student.getSurname());
@@ -275,10 +315,18 @@ public class StudentInfoController implements Initializable {
             email.setText(student.getEmail());
             if (student.getYear() != 0) {
                 gradeLvl.getValueFactory().setValue(student.getYear());            }
+            if (student.getParalel() != 0) {
+                paralel.getValueFactory().setValue(student.getParalel());            }
             try {
                 String relativePath = student.getProfileImg().replace("src/main/resources", "");
-                Image image = new Image(getClass().getResourceAsStream(relativePath));
-                profilePic.setImage(image);
+                InputStream inputStream = getClass().getResourceAsStream(relativePath);
+                if (inputStream != null) {
+                    Image image = new Image(inputStream);
+                    profilePic.setImage(image);
+                } else {
+                    System.out.println("Error loading image: " + student.getProfileImg());
+                }
+                inputStream.close();
             } catch (Exception e)  {
                 System.out.println("Error loading image: " + e.getMessage());
